requirements.txt
fastapi
uvicorn
openai-whisper
requests
numpy
sounddevice
pyaudio
transformers
torch